requirements.txt
astroid==2.3.2
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-dotenv==1.4.2
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
pep8==1.7.1
praw==6.4.0
prawcore==1.0.1
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
update-checker==0.16
urllib3==1.25.6
websocket-client==0.56.0
whitenoise==4.1.4
wrapt==1.11.2
